nuther panic checking .. something is weird here like the serial port interrupts are hooked to the dma interrupts ... it's a bitch just to get the dma irq routines called once ... but if I include a certain printf() in the methods, then it starts looping and flashing like the DMA interrupts are working, though I have my doubts and think it is probably the serial interrupts getting munged.
In addition there seems to be some initialization
that is not being performed correctly, as sometimes
it doesn't happen on a freshly booted rPi until AFTER
I re-run my old program.

Something is smelly in denmark, and I have to
figure it out.  May need to go back to basics
and a single kernel implementation of i2s for
testing.